- NEW: Merged all the changes from the hackathon2016 branch back to the trunk.

+++ src/jhi/brapi/data/LocationDAO.java
@@ -52,7 +52,7 @@ public class LocationDAO
 		location.setName(resultSet.getString("locations.site_name"));
 		location.setCountryCode(resultSet.getString("countries.country_code3"));
 		location.setCountryName(resultSet.getString("countries.country_name"));
-		location.setAltitude(resultSet.getInt("locations.elevation"));
+		location.setAltitude(resultSet.getDouble("locations.elevation"));
 		location.setLatitude(resultSet.getDouble("locations.latitude"));
 		location.setLongitude(resultSet.getDouble("locations.longitude"));
 		location.setLocationDbId(resultSet.getInt("locations.id"));
